Change ScoredArticle to store and retrieve doubles rather than ints
Seems there is a problem converting Double to Integer when using values
from a grouping calculation

+++ src/main/scala/com/persnicketly/model/ScoredArticle.scala
@@ -3,4 +3,4 @@ package com.persnicketly.model
 import com.persnicketly.readability.model.Article
 import org.bson.types.ObjectId
 
-case class ScoredArticle(id: Option[ObjectId], article: Article, favoriteCount: Int, count: Int, score: Double)
+case class ScoredArticle(id: Option[ObjectId], article: Article, favoriteCount: Double, count: Double, score: Double)
